--- _old/src/components/Button/style.styl
@@ -1,109 +1,0 @@
-@import '../../style/required'
-
-@-webkit-keyframes spinning {
-    to {
-        transform(rotate(360deg));
-    }
-}
-
-.g-button
-    border-radius(3px)
-    font-weight 500
-    outline 0
-    border 0
-    cursor pointer
-    text-decoration none
-    display inline-block
-    text-align center
-
-    &.block 
-        width 100%
-        display block
-
-    &--brand-border
-        border 1px solid $brand
-        color $brand
-        background $transparent
-
-    &--white-border 
-        border 1px solid $white
-        color $white
-        background $transparent
-
-    &--black-border 
-        border 1px solid $black
-        color $black
-        background $transparent
-
-    &--brand 
-        border 1px solid $brand
-        color $white
-        background $brand
-
-    &--brand-dark 
-        border 1px solid $brandDark
-        color $white
-        background $brandDark
-
-    &--brand-light
-        border 1px solid $brandLight
-        color $brandDark
-        background $brandLight
-
-    &--blue
-        border 1px solid $blue
-        color $white
-        background $blue
-
-    &--basic-five 
-        border 1px solid $basicFive
-        color $basicTwo
-        background $basicFive
-
-    &--white 
-        color $dark
-        background $white
-
-    &--disabled 
-        opacity 0.5
-        pointer-events(none)
-
-    &--small
-        font-size 13px
-        line-height 16px
-        padding 9px 13px
-
-    &--medium
-        font-size 15px
-        line-height 21px
-        padding 11px 24px
-    
-    &--loading
-        position relative
-        cursor auto
-        pointer-events(none)
-        &:before
-            content '' 
-            position absolute
-            left 0
-            top 0
-            width 100%
-            height 100%
-            background $white
-            opacity 0.4
-            z-index 1
-        &:after
-            content ''
-            display inline-block
-            width 15px
-            height 15px
-            position absolute
-            left 50%
-            margin-left -7.5px
-            top 50%
-            margin-top -7.5px
-            border 2px solid $white
-            border-right-color $dark
-            border-radius(50%)
-            animation(spinning 0.5s linear infinite)
-            z-index 2
