--- _old/src/style/required.styl
@@ -1,2 +1,0 @@
-@import './variables'
-@import './mixins'
