+++ src/assets/style/required.scss
@@ -1,0 +1,2 @@
+@import './variables';
+@import './mixins';
