+++ src/assets/style/variables.scss
@@ -1,0 +1,40 @@
+// COLORS - new
+$brandDark : #00674F;
+$brand : #019875;
+$brandLight : #B0E1D2;
+
+$yellow : #FFE121;
+$orange : #FE8125;
+$red : #FF5F42;
+$blue : #0AAFFF;
+
+$basicOne : #393B46;
+$basicTwo : #4B5169;
+$basicThree : #747A93;
+$basicFour : #B4BAD2;
+$basicFive : #B4BAD2;
+$basicSix : #D1D7DE;
+$basicSeven : #F0F5FB;
+$basicEight : #CED2D4;
+$basicNine : #2B3940;
+
+$white : #ffffff;
+$gray : #E5E5E5;
+$darker : #333333;
+$dark : #444;
+$transparent : transparent;
+$border : #DDDDDD;
+$black : #000000;
+$green : #1AAF76;
+$lightBlack : #2B3940;
+$newRed : #E20D49;
+$borderGray : #D8E1E6;
+
+$nBrand : #1AAF76;
+$nBrandDark : #006155;
+$nSecondary : #FE8125;	
+$nBasicOne : #2B3940;	
+$nBasicTwo : #2B3940;	
+$nBasicFour : #D8E1E6;
+$nBasicFive : #EFF7FA;	
+$nGray : #F6F6F6;
